requirements.txt
glm
PyOpenGL
PyOpenGL_accelerate
pyrr